move metadata.js to steps
