+++ steps/metadata.js
@@ -1,4 +1,3 @@
-#!/usr/bin/env node
 'use strict';
 
 const Request = require('../lib/request');
